fix(onboarding): improve welcome step
